+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Updated surrealql grammar to the latest version
 - Improved live query editor UI
 - Improve console toggle button layout
+- Improve editor focus and tab removal handling
 - Fixed desktop app not being able to connect to http endpoints
 - Fixed combined query result mode incorrectly handling empty and error results
 - Fixed authentication view not updating on session change